Tested and updated nightly/release queries
